+++ relay-general/src/store/geo.rs
@@ -54,6 +54,15 @@ impl GeoIpLookup {
                     .as_ref()
                     .and_then(|city| Some(city.names.as_ref()?.get("en")?.to_string())),
             ),
+            subdivision: Annotated::from(city.subdivisions.as_ref().and_then(|subdivisions| {
+                subdivisions.get(0).and_then(|subdivision| {
+                    subdivision.names.as_ref().and_then(|subdivision_names| {
+                        subdivision_names
+                            .get("en")
+                            .map(|subdivision_name| subdivision_name.to_string())
+                    })
+                })
+            })),
             region: Annotated::from(
                 city.country
                     .as_ref()